requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
crispy-bulma==0.8.0
cryptography==37.0.4
defusedxml==0.7.1
Django~=4.0.6
django-allauth==0.51.0
django-allauth-ui==0.1.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-hashedfilenamestorage==2.4.1
django-widget-tweaks==1.4.12
flake8==5.0.4
gunicorn==20.1.0
boto3==1.24.61
django-storages==1.13.1
idna==3.3
mccabe==0.7.0
oauthlib==3.2.0
pi==0.1.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==5.3.0
dj-database-url==0.5.0
setuptools~=60.2.0

